ref: difference the next and previous multiple
In previous iterations, I was calculating the modulus of the the grade and subtracting it from the grade. This attempt only retrieves the previous multiple. By adding 5 to the previous multiple you get the next multiple. The next step is to subtract the grade from the next multiple to get the remainder. Use this remainder to check if it's less than 3. If yes, then round up the grades